Reorganize the directories and setup the judge logic
