+++ utils/setup.py
@@ -1,0 +1,32 @@
+import time
+from typing import List
+
+import cozmo
+from cozmo.objects import LightCube, LightCube1Id, LightCube2Id, LightCube3Id
+
+
+def setup(robot: cozmo.robot.Robot, cube_color: cozmo.lights.Light) -> (List[LightCube]):
+    """
+    Setup up the cozmo program to run for each computer to use.
+
+    :param robot robot to get cubes for
+    :param cube_color color of this team's cubes
+    """
+
+    # store all of the cube objects in a list
+    robot_cubes: List[LightCube] = [robot.world.get_light_cube(LightCube1Id), robot.world.get_light_cube(LightCube2Id),
+                                    robot.world.get_light_cube(LightCube3Id)]
+
+    # set the colors for robot1's cubes to blue and robot2's to red
+    for cube in range(len(robot_cubes)):
+        robot_cubes[cube].set_lights(cube_color)
+
+    # start the game once the master computer sends out the start message over the network
+    print("Set Cozmo's in position to play.")
+
+    # give a 10 second period for the users to set their robots in their bases and hide their cubes
+    time.sleep(10)
+
+    print("Start playing!")
+
+    return robot_cubes
